Remove key/value parsing from tokenizer
Keeping track of key/value state in the tokenizer is not as helpful as
it seemed, because in practice the parser will need to reimplement
similar state tracking logic in order to produce structured form data.
So instead of trying to track semantic information in the tokenizer,
stick to separating text segments.

We preserve the tracking of runs of value separators ("=") so that the
parser can just treat values as text.